requirements.txt
mutagen
pillow
pygame
pyserial
psutil
